Added m2m signal receivers for Package.data_objects which maintains the Rule instances underneath which are actually used by the poller.
Unfortunately the Django admin does a clear and add on m2m fields every time
you save. Since DataInstances have an FK to the Rule which created them this
would delete all DataInstances for that rule (BAD).

This can easily be avoided in the normal user interface but cannot (without
some heavy admin modification) be avoided in the admin.